feat(part-1: demo): distribute into each component
